[femutils] Move class 'CSRFormatView' in its own file 'CsrFormatMatrixView.h'
